answer most of Em comments
